fix const and light.hbridge
